requirements.txt
PyGithub
httpx